// SFA - FE/packages/connect-force/types/route-types.ts
import { PagedResultStatus } from "connect-force-api-client";


export interface Route {
  uId?: number;
  routeId?: string | null;
  routeName?: string | null;
  description?: string | null;
  startPoint?: string | null;
  endPoint?: string | null;
  distance?: string | null;
  estimateTime?: string | null;
}

export interface RoutePagedResult {
  items?: Array<Route> | null;
  paging?: PagedResultStatus;
}